+++ SampleCode/Android/app/src/main/java/com/netease/meetinglib/demo/view/MeetingSettingsFragment.java
@@ -6,12 +6,16 @@
 package com.netease.meetinglib.demo.view;
 
 import android.os.Bundle;
+import android.text.InputType;
 import android.util.Log;
 
+import androidx.annotation.Nullable;
+import androidx.preference.EditTextPreference;
 import androidx.preference.PreferenceDataStore;
 import androidx.preference.PreferenceFragmentCompat;
 
 import com.netease.meetinglib.demo.R;
+import com.netease.meetinglib.demo.data.MeetingConfigRepository;
 import com.netease.meetinglib.sdk.NEMeetingSDK;
 import com.netease.meetinglib.sdk.NESettingsService;
 
@@ -23,6 +27,8 @@ public class MeetingSettingsFragment extends PreferenceFragmentCompat {
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
         getPreferenceManager().setPreferenceDataStore(new DataStore());
         setPreferencesFromResource(R.xml.meeting_settings, rootKey);
+        ((EditTextPreference) findPreference("join_timeout_millis"))
+                .setOnBindEditTextListener(editText -> editText.setInputType(InputType.TYPE_CLASS_NUMBER));
     }
 
     private static class DataStore extends PreferenceDataStore {
@@ -30,6 +36,32 @@ public class MeetingSettingsFragment extends PreferenceFragmentCompat {
         private final static String ENABLE_SHOW_MEETING_TIME = "enable_show_meeting_time";
         private final static String ENABLE_VIDEO = "enable_video";
         private final static String ENABLE_AUDIO = "enable_audio";
+        private final static String JOIN_TIMEOUT = "join_timeout_millis";
+
+        @Nullable
+        @Override
+        public String getString(String key, @Nullable String defValue) {
+            switch (key) {
+                case JOIN_TIMEOUT:
+                    return String.valueOf(MeetingConfigRepository.INSTANCE.getJoinTimeout());
+            }
+            return super.getString(key, defValue);
+        }
+
+        @Override
+        public void putString(String key, @Nullable String value) {
+            switch (key) {
+                case JOIN_TIMEOUT:
+                    try {
+                        MeetingConfigRepository.INSTANCE.setJoinTimeout(Integer.parseInt(value));
+                    } catch (NumberFormatException e) {
+                        e.printStackTrace();
+                    }
+                    break;
+                default:
+                    super.putString(key, value);
+            }
+        }
 
         @Override
         public boolean getBoolean(String key, boolean defValue) {